+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- Added "no-operation" flag which prevents commands such as `bumpy increment` to persist changes
+
 ## [0.9.0] - 2018-02-09
 
 ### Added